inline json output checkpoint
Some cleanups wrt phred scale reporting; also checkpoint prior to
removing inline json output.

+++ src/Utility.h
@@ -6,11 +6,13 @@
 #include <math.h>
 #include <vector>
 #include <string>
+#include <algorithm>
 
+//long double factorial(int);
 short qualityChar2ShortInt(char c);
 //long double phred2float(int qual);
 long double phred2float(int qual);
-short float2phred(long double prob);
+int float2phred(long double prob);
 // here 'joint' means 'probability that we have a vector entirely composed of true bases'
 short jointQuality(const std::vector<short>& quals);
 short jointQuality(const std::string& qualstr);
